requirements.txt
streamlit
python-dotenv
google-api-python-client
google-auth-oauthlib
langchain_google_genai
langchain